added validation checks for creating/editing group modal
